fix(quality): describe the sport, not one qualifying session
Running the 2024 season through the quality gate showed which of its
thresholds described the one session they were profiled on rather than
Formula 1. Each of these blocked a whole race until it was corrected, and each
correction is recorded in F011's spec with the session and count that forced
it.

- DRS: the enumeration of the seven codes Suzuka Q contained becomes a
  plausible-status-byte check, range(16). Codes 2, 3, 11, 13 and 15 all appear
  in race telemetry and each was added one at a time before it was clear the
  enumeration described the sample, not the channel.
- Pressure: floor 800 -> 700 hPa. Mexico City sits at 2,240 m and reports
  784 hPa; the old floor described sea level and refused the Mexican Grand
  Prix outright.
- delta_t nulls: 0.5 % -> 2 %. A race leaves far more of the grid unshared
  with the reference lap than qualifying does; the Dutch Grand Prix measured
  0.65 %.
- Per-sample channel bounds gain a 1e-3 glitch tolerance. The feed emits
  occasional nonsense -- 49 samples of Miami R report gear 72, two samples of
  British R project more than 100 m from the line -- and global CLAUDE.md 3.1
  says a corrupt record is skipped, not that the batch stops. A test asserts
  the tolerance reaches no Unique, ForeignKey or Invariant rule.
- n_gear: Int8 -> Int16 in the ingest and grid schemas. One sample of the
  Japanese Grand Prix reads 128, exactly one past int8, and the safe cast
  crashed the ingest. The warehouse column was already smallint; the value now
  flows through to be judged by the gate rather than killing the session.

The quality task no longer retries: a contract failure is a deterministic
property of the data just written, and three attempts cost twelve minutes per
failing session during the backfill for the same verdict.

F016's document is re-derived from the re-aligned Suzuka run and its test
follows the session's new slug. The four findings did not move -- T5 16/16 at
+0.122 s, the T8-T9 artefact, reconciliation to 0.0033 s -- but grid zero now
sits in a different place, so the straights row, the sector-1 reconciliation
row and the trace window's coordinates read differently and were regenerated
rather than assumed.

Not amended, deliberately: events.Invariant(event_boundaries_are_ordered)
fails at Spa-Francorchamps, Zandvoort and Lusail. That is F009's segmentation
and the gate is right to refuse it; those sessions stay recorded in the
per-circuit table rather than being let through.

Tests: 844 offline passed.

+++ src/ingest/schemas.py
@@ -105,7 +105,12 @@ CAR_TELEMETRY = ArtefactSchema(
         # D4: Brake is an on/off channel, never a pressure.
         "brake": "boolean",
         "rpm": "float32",
-        "n_gear": "Int8",
+        # Int16, not Int8: the feed emits the occasional garbage gear and one
+        # sample of the 2024 Japanese Grand Prix reads 128 -- exactly one past
+        # int8, which crashed the ingest outright (F015). The warehouse column
+        # is already smallint, and F011 tolerates a sliver of impossible gears
+        # rather than the type refusing to carry them.
+        "n_gear": "Int16",
         "drs": "Int8",
         "source": "string",
     },
